create entrylist api
